modified player progress

+++ Assets/Scripts/PlayerProgress.cs
@@ -13,7 +13,7 @@ public class PlayerProgress : MonoBehaviour {
     public static float dmgTakenMultiplier = 1;
     public static float invicibleTime = 0.1f;
     public static float presentDropChance = 0.2f;
-    public static int walkSpeed= 20;
-    public static int mistleSize = 1;
-    public static int maxHealth = 10;
+    public static float walkSpeed= 20;
+    public static float mistleSize = 1;
+    public static float maxHealth = 10;
 }
